Moved User to models
