+++ submitter.py
@@ -9,9 +9,9 @@ from filemanager import FileManager
 from models.course import Course
 from models.lesson import Lesson
 from models.section import Section
+from models.user import User
 from navigation import prev_step, next_step
 from settings import APP_FOLDER, CLIENT_ID, CLIENT_SECRET, GRAND_TYPE_PASSWORD
-from user import User
 from utils import exit_util
